requirements.txt
django==1.8.5
djangorestframework==3.1.1
django-statsd-mozilla==0.4.0
simplejson==3.8.0
jsonfield==1.0.3
celery[redis]==4.1.0
psycopg2==2.7.3.2
fuzzywuzzy==0.15.1
XlsxWriter==0.9.3
python-dotenv==0.6.5
django-nose==1.4.5